% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{datatable2}
\alias{datatable2}
\title{datatable2 - Datatable with child rows}
\usage{
datatable2(x, vars = NULL, opts = NULL, font.size = "10pt", dom = "fti", ...)
}
\arguments{
\item{x}{A dataframe}

\item{vars}{String vector. The variables you want to put in the details table, rather than
keep in the main table.}

\item{opts}{Additional options for DT::datatable.}

\item{font.size}{String. Default is "10pt".}

\item{dom}{String. What DT::datatable elements to show. Default is 'fti'.}

\item{...}{Additional parameters for DT::datatable.}
}
\value{
A DT::datatable.
}
\description{
Created by Reigo Hendrikson: <http://www.reigo.eu/2018/04/extending-dt-child-row-example/>
}
