% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{num_x}
\alias{num_x}
\title{Numeric x}
\usage{
num_x(x)
}
\arguments{
\item{x}{A vector}
}
\value{
A vector of numbers with NAs removed, if `x` is numeric.
        If `x` is non-numeric or 100% NAs, returns a single NA.
}
\description{
Numeric x
}
